test scheduler.py
